#17 #16 #14 heatmap function. entrypoint and  run: gpu option

+++ run.sh
@@ -6,7 +6,14 @@ out_name="out"
 img_path=$container_path$2
 out_path=$container_path$out_name
 
-echo $img_path
-echo $out_path
-
-docker run --rm -it -v $1:$container_path maevision_tda:pro $img_path $out_path
+# echo $img_path
+# echo $out_path
+if [ $# -eq 3 ]
+then
+    echo "Running app in GPU..."
+    # docker run --rm -it -v $1:$container_path maevision_tda:pro $img_path $out_path
+    echo "Coming soon..."
+else
+    echo "Running app in CPU..."
+    docker run --rm -it -v $1:$container_path maevision_tda:pro $img_path $out_path
+fi
